requirements.txt
llama-index-llms-ollama
llama-index-llms-cohere
llama-index-llms-groq
llama-index-llms-azure-openai
streamlit
pyperclip
plotly